refactor: replace private hardhat submodule with public Foundry contracts
Remove the zaiffer-smart-contracts (hardhat) private submodule and replace
it with a public contracts/ directory powered by Foundry and forge-fhevm.
This unblocks open-sourcing the SDK repo by eliminating its only private
dependency.

- Add contracts/src/ with all Solidity sources (ERC7984, wrapper, factory,
  admin, batcher) adapted for forge-fhevm imports
- Add Deploy.s.sol forge script and deploy-local.sh orchestration
- Add forge-fhevm as a public git submodule under contracts/lib/
- Update Playwright config: anvil webServer, globalSetup for deployment
- Update test fixtures: viem foundry chain, anvil mode, new paths
- Update CI workflow: Foundry toolchain, remove SUBMODULE_TOKEN secret
- Update root package.json scripts: contracts:deps/build/deploy + anvil
