+++ .gitignore
@@ -48,3 +48,9 @@ tools/mdbook/.tmp-mdbook
 # api-extractor temp files
 packages/sdk/temp/
 packages/react-sdk/temp/
+
+# foundry (contracts/)
+contracts/out/
+contracts/cache/
+contracts/broadcast/
+contracts/deployments.json
