+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "hardhat"]
-	path = hardhat
-	url = https://github.com/zama-ai/zaiffer-smart-contracts
+[submodule "contracts/lib/forge-fhevm"]
+	path = contracts/lib/forge-fhevm
+	url = https://github.com/zama-ai/forge-fhevm
