+++ contracts/foundry.toml
@@ -1,0 +1,13 @@
+[profile.default]
+src = "src"
+out = "out"
+libs = ["lib"]
+solc = "0.8.27"
+evm_version = "cancun"
+optimizer = true
+optimizer_runs = 100
+
+fs_permissions = [{ access = "read-write", path = "./" }]
+
+[profile.default.fmt]
+line_length = 120
