+++ contracts/remappings.txt
@@ -1,0 +1,7 @@
+forge-fhevm/=lib/forge-fhevm/src/
+@fhevm/solidity/=lib/forge-fhevm/dependencies/fhevm-0.11.1/library-solidity/
+@openzeppelin/confidential-contracts/=lib/forge-fhevm/dependencies/@openzeppelin-confidential-contracts-6edd293/contracts/
+@openzeppelin/contracts/=lib/forge-fhevm/dependencies/@openzeppelin-contracts-5.1.0/
+@openzeppelin/contracts-upgradeable/=lib/forge-fhevm/dependencies/@openzeppelin-contracts-upgradeable-5.1.0/
+encrypted-types/=lib/forge-fhevm/dependencies/encrypted-types-0.0.4/
+forge-std/=lib/forge-fhevm/dependencies/forge-std-1.14.0/src/
