+++ contracts/src/fhevm/ZamaEthereumConfigUpgradeable.sol
@@ -1,0 +1,20 @@
+// SPDX-License-Identifier: BSD-3-Clause-Clear
+pragma solidity ^0.8.27;
+
+import {FHE} from "@fhevm/solidity/lib/FHE.sol";
+import {ZamaConfig} from "@fhevm/solidity/config/ZamaConfig.sol";
+import {Initializable} from "@openzeppelin/contracts-upgradeable/proxy/utils/Initializable.sol";
+
+abstract contract ZamaEthereumConfigUpgradeable is Initializable {
+    function __ZamaEthereumConfig_init() internal onlyInitializing {
+        __ZamaEthereumConfig_init_unchained();
+    }
+
+    function __ZamaEthereumConfig_init_unchained() internal onlyInitializing {
+        FHE.setCoprocessor(ZamaConfig.getEthereumCoprocessorConfig());
+    }
+
+    function confidentialProtocolId() public view returns (uint256) {
+        return ZamaConfig.getConfidentialProtocolId();
+    }
+}
