+++ package.json
@@ -30,9 +30,10 @@
     "prepare": "husky",
     "submodule:init": "git submodule update --init --recursive",
     "submodule:update": "git submodule update --remote --merge",
-    "hardhat:install": "npm --prefix hardhat install",
-    "hardhat:compile": "npm --prefix hardhat run compile",
-    "hardhat:node": "npm --prefix hardhat run node",
+    "contracts:deps": "cd contracts/lib/forge-fhevm && forge soldeer install",
+    "contracts:build": "cd contracts && forge build",
+    "contracts:deploy": "cd contracts && ./deploy-local.sh",
+    "anvil": "anvil --port 8545 --chain-id 31337 --silent",
     "e2e:test": "pnpm --filter @zama-fhe/playwright e2e:test",
     "e2e:test:ui": "pnpm --filter @zama-fhe/playwright e2e:test:ui",
     "e2e:test:nextjs": "pnpm --filter @zama-fhe/playwright e2e:test:nextjs",
