+++ packages/test-nextjs/src/constants.ts
@@ -1,5 +1,5 @@
 import type { Address } from "@zama-fhe/react-sdk";
-import deployments from "../../../hardhat/deployments.json";
+import deployments from "../../../contracts/deployments.json";
 
 export const CONTRACTS = {
   USDT: deployments.USDT as Address,
